refactor: Extract element assertions
Move the element assertions into a separate class to align the assertions and increase the code coherency.
