+++ testing/camunda-process-test-java/src/main/java/io/camunda/process/test/impl/assertions/ProcessInstanceAssertj.java
@@ -18,10 +18,7 @@ package io.camunda.process.test.impl.assertions;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.fail;
 
-import io.camunda.client.api.search.response.FlowNodeInstance;
-import io.camunda.client.api.search.response.FlowNodeInstanceState;
 import io.camunda.process.test.api.assertions.ElementSelector;
-import io.camunda.process.test.api.assertions.ElementSelectors;
 import io.camunda.process.test.api.assertions.ProcessInstanceAssert;
 import io.camunda.process.test.api.assertions.ProcessInstanceSelector;
 import io.camunda.process.test.api.assertions.ProcessInstanceSelectors;
@@ -29,15 +26,11 @@ import io.camunda.process.test.impl.client.CamundaClientNotFoundException;
 import io.camunda.process.test.impl.client.ProcessInstanceDto;
 import io.camunda.process.test.impl.client.ProcessInstanceState;
 import java.io.IOException;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Predicate;
-import java.util.stream.Collectors;
 import org.assertj.core.api.AbstractAssert;
 import org.awaitility.Awaitility;
 import org.awaitility.core.ConditionTimeoutException;
@@ -48,6 +41,7 @@ public class ProcessInstanceAssertj
     implements ProcessInstanceAssert {
 
   private final CamundaDataSource dataSource;
+  private final ElementAssertj elementAssertj;
   private final VariableAssertj variableAssertj;
   private final String failureMessagePrefix;
 
@@ -63,6 +57,7 @@ public class ProcessInstanceAssertj
     this.dataSource = dataSource;
     failureMessagePrefix =
         String.format("Process instance [%s]", processInstanceSelector.describe());
+    elementAssertj = new ElementAssertj(dataSource, failureMessagePrefix);
     variableAssertj = new VariableAssertj(dataSource, failureMessagePrefix);
   }
 
@@ -86,51 +81,37 @@ public class ProcessInstanceAssertj
 
   @Override
   public ProcessInstanceAssert hasActiveElements(final String... elementIds) {
-    hasElementsInState(
-        asElementIdSelectors(elementIds), FlowNodeInstanceState.ACTIVE, Objects::nonNull);
+    elementAssertj.hasActiveElements(getProcessInstanceKey(), elementIds);
     return this;
   }
 
   @Override
   public ProcessInstanceAssert hasActiveElements(final ElementSelector... elementSelectors) {
-    hasElementsInState(
-        Arrays.asList(elementSelectors), FlowNodeInstanceState.ACTIVE, Objects::nonNull);
+    elementAssertj.hasActiveElements(getProcessInstanceKey(), elementSelectors);
     return this;
   }
 
   @Override
   public ProcessInstanceAssert hasCompletedElements(final String... elementIds) {
-    hasElementsInState(
-        asElementIdSelectors(elementIds),
-        FlowNodeInstanceState.COMPLETED,
-        ProcessInstanceAssertj::isEnded);
+    elementAssertj.hasCompletedElements(getProcessInstanceKey(), elementIds);
     return this;
   }
 
   @Override
   public ProcessInstanceAssert hasCompletedElements(final ElementSelector... elementSelectors) {
-    hasElementsInState(
-        Arrays.asList(elementSelectors),
-        FlowNodeInstanceState.COMPLETED,
-        ProcessInstanceAssertj::isEnded);
+    elementAssertj.hasCompletedElements(getProcessInstanceKey(), elementSelectors);
     return this;
   }
 
   @Override
   public ProcessInstanceAssert hasTerminatedElements(final String... elementIds) {
-    hasElementsInState(
-        asElementIdSelectors(elementIds),
-        FlowNodeInstanceState.TERMINATED,
-        ProcessInstanceAssertj::isEnded);
+    elementAssertj.hasTerminatedElements(getProcessInstanceKey(), elementIds);
     return this;
   }
 
   @Override
   public ProcessInstanceAssert hasTerminatedElements(final ElementSelector... elementSelectors) {
-    hasElementsInState(
-        Arrays.asList(elementSelectors),
-        FlowNodeInstanceState.TERMINATED,
-        ProcessInstanceAssertj::isEnded);
+    elementAssertj.hasTerminatedElements(getProcessInstanceKey(), elementSelectors);
     return this;
   }
 
@@ -150,10 +131,6 @@ public class ProcessInstanceAssertj
   public ProcessInstanceAssert hasVariables(final Map<String, Object> variables) {
     variableAssertj.hasVariables(getProcessInstanceKey(), variables);
     return this;
-  }
-
-  private static List<ElementSelector> asElementIdSelectors(final String[] elementIds) {
-    return Arrays.stream(elementIds).map(ElementSelectors::byId).collect(Collectors.toList());
   }
 
   private void hasProcessInstanceInState(
@@ -219,92 +196,8 @@ public class ProcessInstanceAssertj
     return actualProcessInstance.get().getProcessInstanceKey();
   }
 
-  private void hasElementsInState(
-      final List<ElementSelector> elementSelectors,
-      final FlowNodeInstanceState expectedState,
-      final Predicate<FlowNodeInstance> waitCondition) {
-    final long processInstanceKey = getProcessInstanceKey();
-
-    final AtomicReference<List<FlowNodeInstance>> reference =
-        new AtomicReference<>(Collections.emptyList());
-
-    try {
-      Awaitility.await()
-          .ignoreException(CamundaClientNotFoundException.class)
-          .failFast(
-              () -> {
-                final List<FlowNodeInstance> flowNodeInstances =
-                    reference.get().stream().filter(waitCondition).collect(Collectors.toList());
-                return elementSelectors.stream()
-                    .allMatch(
-                        elementSelector ->
-                            flowNodeInstances.stream().anyMatch(elementSelector::test));
-              })
-          .untilAsserted(
-              () -> {
-                final List<FlowNodeInstance> flowNodeInstances =
-                    dataSource.getFlowNodeInstancesByProcessInstanceKey(processInstanceKey);
-                reference.set(flowNodeInstances);
-
-                final List<FlowNodeInstance> flowNodeInstancesInState =
-                    flowNodeInstances.stream()
-                        .filter(
-                            flowNodeInstance -> flowNodeInstance.getState().equals(expectedState))
-                        .collect(Collectors.toList());
-
-                assertThat(elementSelectors)
-                    .allMatch(
-                        elementSelector ->
-                            flowNodeInstancesInState.stream().anyMatch(elementSelector::test));
-              });
-
-    } catch (final ConditionTimeoutException | TerminalFailureException e) {
-
-      final List<ElementSelector> selectorsNotMatched =
-          elementSelectors.stream()
-              .filter(
-                  elementSelector ->
-                      reference.get().stream()
-                          .noneMatch(
-                              element ->
-                                  elementSelector.test(element)
-                                      && element.getState().equals(expectedState)))
-              .collect(Collectors.toList());
-
-      final String elementsNotInState =
-          selectorsNotMatched.stream()
-              .map(
-                  elementSelector -> {
-                    final FlowNodeInstanceState elementState =
-                        reference.get().stream()
-                            .filter(elementSelector::test)
-                            .findFirst()
-                            .map(FlowNodeInstance::getState)
-                            .orElse(FlowNodeInstanceState.UNKNOWN_ENUM_VALUE);
-
-                    return String.format(
-                        "\t- '%s': %s", elementSelector.describe(), formatState(elementState));
-                  })
-              .collect(Collectors.joining("\n"));
-
-      final String failureMessage =
-          String.format(
-              "%s should have %s elements %s but the following elements were not %s:\n%s",
-              failureMessagePrefix,
-              formatState(expectedState),
-              formatElementSelectors(elementSelectors),
-              formatState(expectedState),
-              elementsNotInState);
-      fail(failureMessage);
-    }
-  }
-
   private static boolean isEnded(final ProcessInstanceDto processInstance) {
     return processInstance != null && processInstance.getEndDate() != null;
-  }
-
-  private static boolean isEnded(final FlowNodeInstance flowNodeInstance) {
-    return flowNodeInstance.getEndDate() != null;
   }
 
   private static String formatState(final ProcessInstanceState state) {
@@ -315,19 +208,5 @@ public class ProcessInstanceAssertj
     } else {
       return state.name().toLowerCase();
     }
-  }
-
-  private static String formatState(final FlowNodeInstanceState state) {
-    if (state == null || state == FlowNodeInstanceState.UNKNOWN_ENUM_VALUE) {
-      return "not activated";
-    } else {
-      return state.name().toLowerCase();
-    }
-  }
-
-  private static String formatElementSelectors(final List<ElementSelector> elementSelectors) {
-    final List<String> selectorList =
-        elementSelectors.stream().map(ElementSelector::describe).collect(Collectors.toList());
-    return AssertFormatUtil.formatNames(selectorList);
   }
 }
